Rename groupby file and give reasonable groupby defaults
