+++ ui/src/utils/groupByTimeSeriesTransform.js
